scripts: Fix shader compilation script
Fix for the case where the files have not yet been generated

+++ scripts/generate_spirv.py
@@ -225,13 +225,17 @@ def write_aggregate_files(shader_data_list, apiname, outdir):
     # For the header file, unless you add a new function/file the header is the same
     # To prevent causing things to recompile again, only write to file if different
     # (ccache was hiding this latency, but MSVC wasn't)
-    with open(out_file_header, "r+") as f:
-        existing_content = f.read()
-        new_content = "\n".join(header_content)
-        if new_content != existing_content:
-            f.seek(0)
+    new_content = "\n".join(header_content)
+    if os.path.exists(out_file_header):
+        with open(out_file_header, "r+") as f:
+            existing_content = f.read()
+            if new_content != existing_content:
+                f.seek(0)
+                f.write(new_content)
+                f.truncate()
+    else:
+        with open(out_file_header, "w") as f:
             f.write(new_content)
-            f.truncate()
 
     with open(out_file_source, "w") as f:
         f.write("\n".join(source_content))
